requirements.txt
# Core utilities
boto3>=1.34,<2
pandas==3.0.2
requests==2.33.1
xlsxwriter==3.2.9

# Google API / auth libraries
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1

# Google Cloud Python libraries
google-cloud-access-approval==1.19.0
google-cloud-aiplatform==1.148.0
google-cloud-api-gateway==1.15.0
google-cloud-api-keys==0.8.0
google-cloud-appengine-admin==1.17.0
google-cloud-artifact-registry==1.21.0
google-cloud-batch==0.21.0
google-cloud-bigquery==3.41.0
google-cloud-bigtable==2.36.0
google-cloud-billing==1.19.0
google-cloud-build==3.36.0
google-cloud-compute==1.47.0
google-cloud-container==2.49.0
google-cloud-dns==0.36.1
google-cloud-firestore==2.27.0
google-cloud-functions==1.23.0
google-cloud-iam==2.22.0
google-cloud-kms==3.12.0
google-cloud-language==2.20.0
google-cloud-logging==3.15.0
google-cloud-monitoring==2.30.0
google-cloud-netapp==0.9.0
google-cloud-orchestration-airflow==1.20.0
google-cloud-pubsub==2.37.0
google-cloud-redis==2.21.0
google-cloud-resource-manager==1.17.0
google-cloud-run==0.16.0
google-cloud-runtimeconfig==0.36.1
google-cloud-scheduler==2.19.0
google-cloud-secret-manager==2.27.0
google-cloud-service-directory==1.17.0
google-cloud-speech==2.38.0
google-cloud-storage==3.10.1
google-cloud-storage-transfer==1.20.0
google-cloud-storageinsights==0.4.0
google-cloud-tasks==2.22.0
google-cloud-translate==3.26.0
google-cloud-videointelligence==2.19.0
google-cloud-vision==3.13.0

# VertexAI
google-genai==1.73.1

# Notes
# Google Groups / Docs APIs:
# https://developers.google.com/docs/api/quickstart/python
#
# IAM Python client libraries:
# https://cloud.google.com/iam/docs/reference/libraries#iam-client-libraries-install-python
#
# Dependency reference for some Google client libs:
# https://developers.google.com/google-ads/api/docs/client-libs/python/dependencies
#
# Optional CLI table rendering:
# pip install gcpwn[table]
# Or (from git clone):
# pip install prettytable==3.17.0
